+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/ir/Nodes.kt
@@ -1359,6 +1359,7 @@ internal data class Rel(
     internal data class Binding(
         @JvmField internal val name: String,
         @JvmField internal val type: CompilerType,
+        @JvmField internal val qualifier: List<Identifier.Simple>? = null,
     ) : PlanNode() {
         public override val children: List<PlanNode> = emptyList()